refactor: split admin and stats hotspots
Extract AdminPage and StatsPage sections into focused components.
Split query helpers by domain while preserving the existing queries barrel API.
